Move bdd_to_term into bdd.rs instead of finite.rs
